Create post router and update database schemas

+++ src/database/specifications/0.0.3.schema.js
@@ -12,7 +12,9 @@ const deploy = async knex => {
   }
 }
 const rollback = async knex => {
-  await knex.schema.dropTable('tokens')
+  if (await knex.schema.hasTable('tokens')) {
+    await knex.schema.dropTable('tokens')
+  }
 }
 
 export {
